The input is reverted back to it's usual state a millisecond after form submit
During form submit, it needs to have the 24hr time value in the input field. Most of the time, it needs to display the 12hr time value though.

To avoid form submit issues, it quickly switches the value to 24hr time then switches back to 12hr time after a milisecond. This should be enough time for the form to submit the real value before it swaps back to the 12hr value.