Adding a second "$" in front of $inputs
This is a coding practice I've picked up to represent the variable contains a list of elements rather than a single element.

I also normally put an underscore in front to represent "not-jQuery" but I use $input throughout the documentation and I don't think _$input all throughout the documentation would look very nice. I also want the method used to describe inputs to be consistent.